rename sync client class name
